clean-up fetch helpers / responses; fix delete & reset password actions

+++ src/routes/api/deleteUser.js
@@ -1,9 +1,8 @@
-import * as auth from '$lib/apis/auth-api';
-import * as db from '$lib/apis/db-api';
+import * as auth from '$lib/apis/auth-api-methods';
+import * as db from '$lib/apis/db-api-methods';
+import * as cookie from '$lib/utils/cookies';
 import loadStripe from 'stripe';
-import { parseJwt } from '$lib/utils/helpers';
-import verifyCurrentToken from '$lib/utils/verifyCurrentToken';
-import errorResponse from '$lib/utils/errorResponse';
+import { serverResponse } from '$lib/utils/helpers';
 
 export async function post(request) {
 
@@ -12,47 +11,64 @@ export async function post(request) {
     const { password } = request.body;
 
     // authorize request
-    const accessTokenCheck = await verifyCurrentToken(token); // validate user session
+    const accessTokenCheck = await auth.verifyCurrentToken(token); // validate user session
+    if(!accessTokenCheck.ok) throw { statusMessage: 'error', errorMessage: 'Unauthorized Session' };
+
     const authCheck = await auth.loginUser(user.email, password); // validate user entered password
-    const authCheckEmail = authCheck.body.access_token ? parseJwt(authCheck.body.access_token).email : null; // retrieve validated email
-    const emailMatch = (authCheckEmail === user.email); // compare returned email to user email in session
+    if(authCheck.error) throw { statusMessage: authCheck.statusMessage, errorMessage: 'Please double-check current password.' };
+
+    const emailMatch = (accessTokenCheck.email === user.email); // compare returned email to user email in session
+    if(!emailMatch) throw { statusMessage: 'error', errorMessage: 'Unauthorized User' };
 
     if(accessTokenCheck.ok && authCheck && emailMatch) {
 
-      // attempt to delete user from Netlify Identity
+      // Attempt to delete user from Netlify Identity
       // -------------------------------------------------------------------------------------------
-      const deleteIdentityUserRequest = await auth.deleteUser(token);
+      const deleteIdentityUser = await auth.deleteUser(token);
       // if attempt fails report error on server and throw error for client
-      if (deleteIdentityUserRequest.status !== 204) {
-        console.log(Date.now(), '💥 DELETE IDENTITY USER Unsuccessful - ', deleteIdentityUserRequest);
-        throw { status: deleteIdentityUserRequest.status, message: 'Unable to delete account - please try again.', };
-      } // else, continue
+      if (deleteIdentityUser.body?.error) {
+        console.log(new Date().toISOString(), '💥 DELETE IDENTITY USER unsuccessful :', deleteIdentityUser);
+        throw { statusMessage: `error (${deleteIdentityUser.status})`, errorMessage: 'Unable to delete account - please try again.', };
+      }
       
-      // attempt to delete customer from Stripe
-      // note: this must be done before User Database deletion to have Stripe ID avaialble
+      // Attempt to delete customer from Stripe
+      // note: this must be done *before* user database deletion to have Stripe ID available
       // -------------------------------------------------------------------------------------------
-      const netlifyId = await accessTokenCheck.body.id;
+      const netlifyId = await accessTokenCheck.id;
       const stripeId = await db.getStripeId(netlifyId);
       const stripe = await loadStripe(process.env['STRIPE_SECRET_KEY']);
       const deleteStripeCustomer = await stripe.customers.del(stripeId); // https://stripe.com/docs/api/customers/delete
       // if attempt fails report error on server
-      if (deleteStripeCustomer && deleteStripeCustomer.deleted !== true) console.log(Date.now(), '💥 DELETE STRIPE CUSTOMER Unsuccessful - ', deleteStripeCustomer);
+      if (deleteStripeCustomer && deleteStripeCustomer.deleted !== true) {
+        console.log(new Date().toISOString(), '💥 DELETE STRIPE CUSTOMER unsuccessful :', deleteStripeCustomer)
+      };
 
-      // attempt to delete user from User Database
+      // Attempt to delete user from User Database
       // -------------------------------------------------------------------------------------------
-      await db.deleteUser(netlifyId); // if attempt fails db.deleteUser() method will report error on server
+      await db.deleteUser(netlifyId);
+      // if attempt fails db.deleteUser() method will report error on server
 
-      // attempt to clear auth cookies
+      // Attempt to clear auth cookies
       // -------------------------------------------------------------------------------------------
       const body = {};
-      return auth.deleteIdentityCookies(deleteIdentityUserRequest.status, body);
+      return cookie.deleteIdentityCookies(body);
 
-    } else {
-      throw { status: authCheck.status || 401, message: authCheck.body.error || 'Unauthorized request.', };
     }
 
   } catch (error) {
-    return errorResponse(error, 'deleteUser');
+    let { statusMessage, errorMessage } = error;
+    if (!errorMessage) console.log(new Date().toISOString(), "💥 'deleteUser' endpoint unsuccessful : error.message :", error.message);
+
+    if (errorMessage?.toLowerCase().includes('bad request')) {
+      errorMessage = 'We encountered a system error - please try again.'
+    }
+    if (errorMessage?.toLowerCase().includes('user not found') || errorMessage?.toLowerCase().includes('no user found')) {
+      errorMessage = 'Unauthorized, please double-check password.'
+    }
+    return serverResponse(200, false, {
+      statusMessage: statusMessage || 'error',
+      error: errorMessage || 'Unsuccessful Request',
+    });
   }
 
 }
